Add easy pack change in zone picking (shopfloor)
On the zone picking scenario it is possible to select a move line
by scanning the pack assigned to the move line.

But for some situation (block storage) where a location contains
multiple pack of the same type it may be difficult to scan the exact
pack, and much easier to scan an identical one in front of the user.

This change allows for this so when scanning a pack that can replace a
pack in the existing move line, the user is offered the option to do so,
by scanning the same pack a 2nd time.